requirements.txt
arrow==0.4.2
ics>=0.7
pytz>=2020.4
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
discord.py>=1.5.1
Pillow>=8.0.1